backend/pyproject.toml
[project]
name = "cpr-assist-backend"
version = "0.1.0"
description = "CPR Assist FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "pymongo>=4.8.0",
  "email-validator>=2.2.0",
  "torch>=2.0",
  "torchvision>=0.15",
  "numpy>=1.24",
  "opencv-python>=4.8",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "ultralytics>=8.0",
  "mediapipe>=0.10.9",
  "scikit-learn>=1.3",
  "joblib>=1.3",
  "xgboost>=2.0",
  "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
markers = [
  "integration: tests that require cpr_ml + vision dependencies",
]
